+++ src/modal/DialogBox.jsx
@@ -6,14 +6,17 @@
  * 
  * Change Log
  * =========================================================
- * Date                Version     Author    Description
+ * Date         Version   Author    Description
  * =========================================================
- * 2025/11/27    1.0.0         ITA         Genesis. 
+ * 2025/11/27   1.0.0     ITA       Genesis.
+ * 2025/11/30   1.0.1     ITA       When this component was imported from its npm published  package (dialogboxjs) into a 
+ *                                  vite based project, it complained about a missing PropTypes default import in prop-types.
+ *                                  Vite required explicit non-default import of prop-types object.
  */
 /** File: ./src/modal/DialogBox.jsx */
 import { useEffect, useState } from 'react';
 import Modal from 'react-modal';
-import PropTypes from 'prop-types';
+import { node, object } from 'prop-types';
 import { useDialogBox } from "./DialogBoxProvider"
 
 // Ensure react-modal accessibility works
@@ -65,6 +68,6 @@ export default function DialogBox({ styles = null, children }) {
 }
 
 DialogBox.propTypes = {
-    styles: PropTypes.object,
-    children: PropTypes.node.isRequired,
+    styles: object,
+    children: node.isRequired,
 };